+++ insteon_mqtt/config.py
@@ -9,6 +9,7 @@ __doc__ = """Configuration file utilties
 
 #===========================================================================
 import functools
+import os.path
 import yaml
 from . import device
 
@@ -31,8 +32,8 @@ devices = {
 def load(path):
     """TODO: doc
     """
-    config = yaml.load(open("config.yaml").read())
-    return config
+    with open(path, "r") as f:
+        return yaml.load(f, Loader)
 
 
 #===========================================================================
@@ -68,5 +69,55 @@ def find(name):
                         "%s." % (name, devices.keys()))
 
     return dev
+
+#===========================================================================
+
+
+# YAML multi-file loading helper.  Original code is from here:
+# https://davidchall.github.io/yaml-includes.html (with no license so
+# I'm assuming it's in the public domain).
+class Loader(yaml.Loader):
+    def __init__(self, file):
+        """Constructor
+
+        Args:
+          file:  (file) File like object to read from.
+        """
+        yaml.Loader.add_constructor('!include', Loader.include)
+
+        super().__init__(file)
+        self._base_dir = os.path.split(file.name)[0]
+
+    #-----------------------------------------------------------------------
+    def include(self, node):
+        """!include file command.  Supports:
+
+        foo: !include file.yaml
+        foo: !include [file1.yaml, file2.yaml]
+        """
+        # input is a single file to load.
+        if isinstance(node, yaml.ScalarNode):
+            include_file = self.construct_scalar(node)
+            return self._load_file(include_file)
+
+        # input is a list of files to load.
+        elif isinstance(node, yaml.SequenceNode):
+            result = []
+            for include_file in self.construct_sequence(node):
+                result += self._load_file(include_file)
+            return result
+
+        else:
+            msg = "Error:: unrecognized node type in !include statement: " \
+                  "%s" % str(node)
+            raise yaml.constructor.ConstructorError(msg)
+
+    #-----------------------------------------------------------------------
+    def _load_file(self, filename):
+        """Read the requested file.
+        """
+        path = os.path.join(self._base_dir, filename)
+        with open(path, 'r') as f:
+            return yaml.load(f, Loader)
 
 #===========================================================================